refactor: separate defineMacro.js from macros.js
Allows defineMacro to be called from other places, in particular
removing cycles from the dependency graph.
